moving mvn and gradle tests to new test files
